src/main/java/com/github/christianj98/primarycustomerbase/service/AddressServiceImpl.java
package com.github.christianj98.primarycustomerbase.service;

import com.github.christianj98.primarycustomerbase.dto.AddressDto;
import com.github.christianj98.primarycustomerbase.mapper.AddressMapperService;
import com.github.christianj98.primarycustomerbase.repository.AddressRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class AddressServiceImpl implements AddressService {
    private final AddressRepository addressRepository;
    private final AddressMapperService addressMapperService;

    public List<AddressDto> findAll() {
        return addressMapperService.mapFrom(addressRepository.findAll());
    }
}
